MenuPage Adjustments
Fixed no longer being able to see where you are in a menu page.

Added: Now you can click on the home text to go back to the home page. It displays to the user with Home > YourPage.

Fixed, capitalized headers in database.js.
Improved the headers in the menu.