Added coupons + Cart fixes
Coupons now work functionally.
Coupons will show the user why theirs is not applied, and that switching coupons will remove their current coupon.
Coupons modifications will be applied directly to the cart. If they do not change anything, then a tip is provided to explain why.
Fixed the cart not closing when clicking off of it in a menu page.