Lots of changes -> See description
Fixed error in checkout page with MainPage import.
Added a notification when adding and removing items from the cart.
Adjusted the allignment in the cart for the prices.
Added a button to 'remove an item' from the cart from a food item menu. This will only be displayed if the item is already in the cart.
Added special instructions feature. The feature is now implemented and is displayed correctly on the cart.
Special instructions are treated as a unique item, with the exception of item removal from an item menu.
Various fixes with problem when removing items leaving commas as the items saved from the cart.
Fixed search bar not working when pressing 'enter'. Capitilization is ignored.